refactor: move check_session_expiration to its own file
