+++ src/main.py
@@ -1,20 +1,13 @@
 from quart import Quart, jsonify, request
 from quart_cors import cors
-from datetime import datetime, timedelta
 from routes.debug_routes import debug_routes
 from routes.login_route import login_route
 from routes.user_route import user_route
 from routes.chat_route import chat_route
-from db import Session as DBSession
-import asyncio
-import telebot
 from bot import bot
 from telebot import types
-from config import Config
-from models import Session as SessionModel
-from services.session_service import delete_session
-from telethon.sessions import StringSession
-from telethon.sync import TelegramClient
+from services.session_expiration import check_session_expiration
+
 
 app = Quart(__name__)
 app = cors(app, allow_origin="*")
@@ -23,49 +16,6 @@ app.register_blueprint(debug_routes)
 app.register_blueprint(login_route)
 app.register_blueprint(user_route)
 app.register_blueprint(chat_route)
-
-
-async def check_session_expiration():
-    while True:
-        db_session = DBSession()
-        try:
-            all_sessions = db_session.query(SessionModel).all()
-
-            for user_session in all_sessions:
-                time_difference = datetime.now() - user_session.creation_date
-                # if time_difference >= timedelta(minutes=4):
-                if time_difference >= timedelta(
-                    minutes=Config.SESSION_EXPIRATION_MINUTES
-                ):
-                    print(f"Session for {user_session.phone_number} has expired.")
-                    if user_session.is_logged:
-                        try:
-                            client = TelegramClient(
-                                StringSession(my_session.id),
-                                Config.API_ID,
-                                Config.API_HASH,
-                            )
-                            client = TelegramClient(
-                                StringSession(user_session.id),
-                                Config.API_ID,
-                                Config.API_HASH,
-                            )
-                            await client.connect()
-                            if await client.is_user_authorized():
-                                await client.log_out()
-                        except Exception as e:
-                            print(f"Error in log_out(): {str(e)}")
-                        finally:
-                            await client.disconnect()
-                    await delete_session(user_session.phone_number, None)
-                else:
-                    print(f"Session for: {user_session.phone_number} is active")
-        except Exception as e:
-            print(f"Database error: {str(e)}")
-        finally:
-            db_session.close()
-        # Wait for the specified interval before checking again
-        await asyncio.sleep(Config.CHECK_INTERVAL_SECONDS)
 
 
 @app.before_serving
